swml/parsing/inline-classes-1.dat

#data
aa[YOKO()[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw yoko>
|         "bbb"
|       "o"

#data
aa[YOKO(foo bar)[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw yoko>
|         class="foo bar"
|         "bbb"
|       "o"

#data
aa[YOKO(  foo foo bar #baz .abc #abc .foob x )[bbb]]o
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "aa"
|       <sw yoko>
|         class="foo foo bar x"
|         id="baz"
|         itemprop="abc foob"
|         "bbb"
|       "o"
